% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{timeSeriesQC-package}
\alias{timeSeriesQC}
\alias{timeSeriesQC-package}
\title{Internal imports and globals}
\description{
Import lightweight symbols and declare data-masked variables to keep R CMD check quiet.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yryrena/timeSeriesQC}
  \item Report bugs at \url{https://github.com/yryrena/timeSeriesQC/issues}
}

}
\author{
\strong{Maintainer}: Ruiyi Yang \email{yryrena@gmail.com}

}
\keyword{internal}
