% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest.R
\name{tsqc_suggest}
\alias{tsqc_suggest}
\title{Suggest transformations: log/Box-Cox, differencing, seasonal adjustment}
\usage{
tsqc_suggest(scan_obj)
}
\arguments{
\item{scan_obj}{tsqc_scan}
}
\description{
Heuristics per series:
\itemize{
\item log vs shifted_log
\item difference? (based on lag-1 ACF)
\item seasonal adjustment recommended? (if STL seasonal strength > 0.2)
}
}
